TON of change, as always

+++ src/lib/letsencrypt.ts
@@ -3,15 +3,11 @@ import { forceSSLOffApplication, forceSSLOnApplication, getNextTransactionId, re
 import { asyncExecShell, getEngine } from "./common"
 import * as db from '$lib/database';
 
-export async function letsEncrypt(data) {
-    const { destinationDocker, domain, forceSSLChanged, isCoolify = false, id } = data
+export async function letsEncrypt({ domain, isCoolify = false, id }) {
+    const { destinationDocker, destinationDockerId } = await db.prisma.application.findUnique({ where: { id }, include: { destinationDocker: true } })
     try {
         if (dev) {
-            if (forceSSLChanged) {
-                return await forceSSLOnApplication({ domain })
-            } else {
-                return await forceSSLOffApplication({ domain })
-            }
+            return await forceSSLOnApplication({ domain })
         } else {
             if (isCoolify) {
                 await asyncExecShell(`docker run --rm --name certbot -p 9080:9080 -v "coolify-letsencrypt:/etc/letsencrypt" certbot/certbot --logs-dir /etc/letsencrypt/logs certonly --standalone --preferred-challenges http --http-01-address 0.0.0.0 --http-01-port 9080 -d ${domain} --agree-tos --non-interactive --register-unsafely-without-email`)
@@ -22,31 +18,27 @@ export async function letsEncrypt(data) {
                 return
             }
             // Set SSL with Let's encrypt
-            if (destinationDocker) {
-                if (forceSSLChanged) {
-                    const host = getEngine(destinationDocker.engine)
-                    // saveBuildLog({ line: 'Requesting SSL cert.', buildId })
-                    await asyncExecShell(`DOCKER_HOST=${host} docker run --rm --name certbot -p 9080:9080 -v "coolify-letsencrypt:/etc/letsencrypt" certbot/certbot --logs-dir /etc/letsencrypt/logs certonly --standalone --preferred-challenges http --http-01-address 0.0.0.0 --http-01-port 9080 -d ${domain} --agree-tos --non-interactive --register-unsafely-without-email`)
-                    // saveBuildLog({ line: 'SSL cert requested successfully!', buildId })
-                    // saveBuildLog({ line: 'Parsing SSL cert.', buildId })
-                    const { stderr } = await asyncExecShell(`DOCKER_HOST=${host} docker run --rm --name bash -v "coolify-letsencrypt:/etc/letsencrypt" -v "coolify-ssl-certs:/app/ssl" alpine:latest cat /etc/letsencrypt/live/${domain}/fullchain.pem /etc/letsencrypt/live/${domain}/privkey.pem > /app/ssl/${domain}.pem`)
-                    if (stderr) throw new Error(stderr)
+            if (destinationDockerId && destinationDocker) {
+                const host = getEngine(destinationDocker.engine)
+                // saveBuildLog({ line: 'Requesting SSL cert.', buildId })
+                await asyncExecShell(`DOCKER_HOST=${host} docker run --rm --name certbot -p 9080:9080 -v "coolify-letsencrypt:/etc/letsencrypt" certbot/certbot --logs-dir /etc/letsencrypt/logs certonly --standalone --preferred-challenges http --http-01-address 0.0.0.0 --http-01-port 9080 -d ${domain} --agree-tos --non-interactive --register-unsafely-without-email`)
+                // saveBuildLog({ line: 'SSL cert requested successfully!', buildId })
+                // saveBuildLog({ line: 'Parsing SSL cert.', buildId })
+                const { stderr } = await asyncExecShell(`DOCKER_HOST=${host} docker run --rm --name bash -v "coolify-letsencrypt:/etc/letsencrypt" -v "coolify-ssl-certs:/app/ssl" alpine:latest cat /etc/letsencrypt/live/${domain}/fullchain.pem /etc/letsencrypt/live/${domain}/privkey.pem > /app/ssl/${domain}.pem`)
+                if (stderr) throw new Error(stderr)
 
-                    // saveBuildLog({ line: 'SSL cert parsed.', buildId })
-                    // saveBuildLog({ line: 'Reloading Haproxy', buildId })
-                    await forceSSLOnApplication({ domain })
+                // saveBuildLog({ line: 'SSL cert parsed.', buildId })
+                // saveBuildLog({ line: 'Reloading Haproxy', buildId })
+                await forceSSLOnApplication({ domain })
 
-                } else {
-                    await forceSSLOffApplication({ domain })
-                }
                 await reloadConfiguration()
             }
 
         }
     } catch (err) {
-        if (!dev && !isCoolify) {
-            await db.prisma.applicationSettings.updateMany({ where: { applicationId: id }, data: { forceSSL: false } })
-        }
+        // if (!dev && !isCoolify) {
+        //     await db.prisma.applicationSettings.updateMany({ where: { applicationId: id }, data: { forceSSL: false } })
+        // }
         throw err
     }
 }
